Better Type IDs

+++ src/main/kotlin/xyz/xenondevs/nova/serialization/cbf/Element.kt
@@ -16,7 +16,7 @@ import kotlin.reflect.KFunction
 
 interface Element {
     
-    fun getTypeId(): Byte
+    fun getTypeId(): Int
     
     fun write(buf: ByteBuf)
